3hw.5й - поправки ментора
