+++ Schemas/Schemas.js
@@ -1,0 +1,34 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+  name: Joi.string().min(3).max(30).required(),
+  email: Joi.string()
+    .email({
+      minDomainSegments: 2,
+      tlds: { allow: ["com", "net"] },
+    })
+    .required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
+
+const updateSchema = Joi.object({
+  name: Joi.string().min(3).max(30),
+  email: Joi.string().email({
+    minDomainSegments: 2,
+    tlds: { allow: ["com", "net"] },
+  }),
+  phone: Joi.string(),
+  favorite: Joi.boolean(),
+}).min(1);
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+const schemas = {
+  addSchema,
+  updateSchema,
+  updateFavoriteSchema,
+};
+
+module.exports = schemas;
