+++ middlewares/isValidId.js
@@ -1,7 +1,7 @@
 const { isValidObjectId } = require("mongoose");
 const { BadRequest } = require("http-errors");
 
-const isVakidId = (req, res, next) => {
+const isValidId = (req, res, next) => {
   const { contactId } = req.params;
   const isCorrectId = isValidObjectId(contactId);
   if (!isCorrectId) {
@@ -13,4 +13,4 @@ const isVakidId = (req, res, next) => {
   next();
 };
 
-module.exports = isVakidId;
+module.exports = isValidId;
