+++ models/contacts.js
@@ -1,6 +1,5 @@
 const { Schema, model } = require("mongoose");
 const handleSchemaErrors = require("../helpers/handleSchemaErrors");
-const Joi = require("joi");
 
 const contactsSchema = new Schema(
   {
@@ -27,37 +26,6 @@ const contactsSchema = new Schema(
 
 contactsSchema.post("save", handleSchemaErrors);
 
-const addSchema = Joi.object({
-  name: Joi.string().min(3).max(30).required(),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net"] },
-    })
-    .required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-
-const updateSchema = Joi.object({
-  name: Joi.string().min(3).max(30),
-  email: Joi.string().email({
-    minDomainSegments: 2,
-    tlds: { allow: ["com", "net"] },
-  }),
-  phone: Joi.string().min(1),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-const schemas = {
-  addSchema,
-  updateSchema,
-  updateFavoriteSchema,
-};
-
 const Contact = model("contact", contactsSchema);
 
-module.exports = { Contact, schemas };
+module.exports = Contact;
